Remove require_tree . from Asset Pipeline
require_tree . is probably too broad for our purposes. It processes ALL the scripts in the directory tree, but the order is undefined (a problem if you require a ‘core’ component to be loaded first). 
You may also end up with duplicated functions if you’re referencing a library which comes bundled with say, JQuery, and you’re also referencing JQuery separately.